edugpt/requirements.txt
accelerate==0.24.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.6.0
async-timeout==4.0.3
async-tio==1.3.2
attrs==23.1.0
BingImageCreator==0.5.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
curl-cffi==0.5.7
dataclasses-json==0.5.14
datasets==2.17.0
dill==0.3.8
discord.py==2.3.0
EdgeGPT==0.10.13
exceptiongroup==1.1.3
filelock==3.13.1
FlagEmbedding==1.2.5
frozenlist==1.4.0
fsspec==2023.10.0
GoogleBard==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.277
langsmith==0.0.30
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==0.27.8
OpenAIAuth==3.0.0
outcome==1.2.0
packaging==23.1
pandas==2.2.0
pillow==10.2.0
prompt-toolkit==3.0.39
psutil==5.9.6
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
revChatGPT==6.3.3
rich==13.5.2
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.9.1
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.4.0
tls-client==0.2.1
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.37.2
trio==0.22.2
trio-websocket==0.10.3
triton==2.1.0
typing-inspect==0.9.0
typing-extensions==4.7.1
tzdata==2024.1
undetected-chromedriver==3.4.7
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
